requirements.txt
adafruit_circuitpython_debouncer
adafruit_circuitpython_macropad
adafruit_circuitpython_pixelbuf
adafruit_circuitpython_simple_text_display
adafruit_circuitpython_ticks
neopixel
adafruit_circuitpython_bitmap_font
adafruit_circuitpython_display_text
adafruit_circuitpython_hid
adafruit_circuitpython_midi
adafruit_circuitpython_display_shapes
adafruit_circuitpython_displayio_layout
adafruit_circuitpython_imageload
